% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters.R
\name{create_parset}
\alias{create_parset}
\title{Create parameter set from a template and random numbers'}
\usage{
create_parset(sample, parameters, .fun = NULL)
}
\arguments{
\item{sample}{A numeric vector with length n, where n is the number of \code{NA}s in the parameters list}

\item{parameters}{A list of lists (or data.frames) with \code{NA}s where a parameter should be sampled. Each atomic element can have an arbitratry numer of \code{NA} values and attributes.}

\item{.fun}{Optional. A function to be applied to each \code{NA} in the atomic vector. Its first argument is the nth sample correspoding to the nth \code{NA} value in th whole list. Attributes stored on that vector will be passed as further arguments to the function.}
}
\value{
A list of lists of the properties length as template
}
\description{
From random uniform samples to paramaters
 - recursivley iterate over a list of lists (or data frames)
 - each element is an atomic vector with possible attributes
 - if a vector has attributes, on every NA in that vector the rescale function is called. The arguments for that function are
   - a random number [0,1] that is fetched from a vector of random numbers with length nvar which is the number of NAs in the list of lists vector of random numbers
    - the attributes belonging to the atomic vector
}
