% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp.R
\name{interception}
\alias{interception}
\title{Title}
\usage{
interception(p, evaporation, lai, a = 0.025, k = 0.39)
}
\arguments{
\item{p}{precipitation}

\item{evaporation}{evaporation}

\item{lai}{lai}

\item{a}{a}

\item{k}{k}
}
\value{
data.frame
}
\description{
Title
}
