% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp.R
\name{split_etp}
\alias{split_etp}
\title{Split ET_p into Evaporation and Transpiration}
\usage{
split_etp(etp, lai, k = 0.39)
}
\arguments{
\item{etp}{evapotranspiratio}

\item{lai}{Leaf area index}

\item{k}{extinction coefficient}
}
\value{
data.frame with columns evaporation and transpiration
}
\description{
Based on Beer's law , this function splits potential evapotranspiration into
potential transpiration and potential evaporation depending on Leaf Area Index
(LAI) and the extinction factor k.
}
