% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_profile}
\alias{write_profile}
\title{Write soil profile information to file}
\usage{
write_profile(pars, path = ".", file_name = "PROFILE.DAT")
}
\description{
Write soil profile information to file
}
