% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_selector}
\alias{write_selector}
\title{Write file SELECTOR.IN for HYDRUS-1D}
\usage{
write_selector(pars, path = ".", file_name = "SELECTOR.IN")
}
\arguments{
\item{pars}{List of parameters}

\item{path}{Path to folder where SELECTOR.IN should be created. If missing, the
function writes to stdout.}

\item{file_name}{Name of file}
}
\value{
List of parameters
}
\description{
Write file SELECTOR.IN for HYDRUS-1D
}
