% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output}
\alias{read_output}
\alias{read_output}
\alias{read_t_level}
\title{Read from HYDRUS-1D output files}
\usage{
read_output(file_path, skip)

read_t_level(path)
}
\arguments{
\item{file_path}{Path to *.out file}

\item{skip}{skip number of lines on input}

\item{path}{Path to Output directory}

\item{skip}{Skip number of lines on input}
}
\value{
data.frame

data.frame
}
\description{
Read from HYDRUS-1D output files
}
