% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output}
\alias{read_output}
\alias{read_t_level}
\alias{read_a_level}
\alias{read_nod_inf}
\alias{read_obs_node}
\title{Read HYDRUS-1D output files into R}
\usage{
read_output(project, ...)

read_t_level(file_path, skip = 6L)

read_a_level(file_path, skip = 2L)

read_nod_inf(file_path, colnames_row = 11L)

read_obs_node(file_path, header = 11L)
}
\arguments{
\item{project}{Either a h1d_run object or a path to a file (single string)}

\item{file_path}{Path to T_level output file}

\item{skip}{skip number of lines on input}

\item{header}{Number of lines with header information before the actual data}
}
\value{
list of data frames

data.frame

data.frame

list of data.frames

data.frame
}
\description{
Read HYDRUS-1D output files into R
}
