% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hydrus.R
\name{run_h1d}
\alias{run_h1d}
\title{Execute H1D from within R}
\usage{
run_h1d(h1d_exec, project_path)
}
\arguments{
\item{h1d_exec}{Path to H1D_CALC executable}

\item{project_path}{Path to HYDRUS project folder}
}
\value{
list
}
\description{
Execute H1D from within R
}
