% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_h1d.R
\name{run_h1d}
\alias{run_h1d}
\title{Execute H1D from within R}
\usage{
run_h1d(project, h1d_exec = "h1d_calc", async = FALSE)
}
\arguments{
\item{project}{Path to HYDRUS project folder}

\item{h1d_exec}{Path to H1D_CALC executable}
}
\value{
list
}
\description{
Execute H1D from within R
}
