% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_selector}
\alias{write_selector}
\title{Write file SELECTOR.IN for HYDRUS-1D}
\usage{
write_selector(pars = list(), folder, ...)
}
\arguments{
\item{pars}{List of parameters}

\item{folder}{Path to folder where SELECTOR.IN should be created. If missing, the
function writes to stdout.}

\item{...}{Additional parmaters}
}
\value{
Listt of parameters
}
\description{
Write file SELECTOR.IN for HYDRUS-1D
}
