% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_selector}
\alias{write_selector}
\title{Write file SELECTOR.IN for HYDRUS-1D}
\usage{
write_selector(path, ..., options = list(Head = "run from R", LUnit = "cm",
  TUnit = "days", MUnit = "mmol", lWat = TRUE, lChem = TRUE, lTemp = TRUE, lSink
  = TRUE, lRoot = FALSE, lShort = FALSE, lWDep = FALSE, lScreen = TRUE,
  lVariabBC = TRUE, lEquil = TRUE, lSnow = TRUE, lMeteo = FALSE, lVapor = FALSE,
  lActRSU = FALSE, lFlux = FALSE, lIsotope = FALSE, NMat = 1, NLay = 1, CosAlf =
  1, MaxIt = 20, TolTh = 0.001, TolH = 1, TopInF = TRUE, WLayer = TRUE, KodTop =
  -1, lInitW = FALSE, BotInF = FALSE, qGWLF = FALSE, FreeD = TRUE, SeepF =
  FALSE,      KodBot = -1, qDrain = FALSE, hSeep = 0, hTab1 = 1e-06, hTabN =
  1e+06, iModel = 0, iHyst = 0, iKappa = 1, hydraulics = data.frame(thr = 0.08,
  ths = 0.45, Alfa = 0.036, n = 1.56, Ks = 25, l = 0.5), dt = 0.01, dtMin =
  1e-07, dtMax = 1, dMul = 1.3, dMul2 = 0.7, ItMin = 3, ItMax = 7, MPL = 10,
  tInit = 0, tMax = 1, TPrint = NA, lPrintD = TRUE, nPrStep = 1, tPrintInt = 1,
  lEnter = FALSE, heat = data.frame(Qn = 0.4, Qo = 0, Disper = 5, B1 =
  1.47054e+16, B2 = -1.5518e+17, B3 = 3.16617e+17, Cn = 1.43327e+14,          Co
  = 1.8737e+14, Cw = 3.12035e+14)), file_version = 4L)
}
\arguments{
\item{path}{}
}
\description{
Write file SELECTOR.IN for HYDRUS-1D
}
